Add FloatingPoint Instructions (excluding pipeline for now)

+++ src/main/java/thb/fbi/leguan/controller/MemoryController.java
@@ -45,13 +45,13 @@ public class MemoryController implements MemoryObserver {
     TableColumn<Map.Entry<Long, Long>, String> contentColumn;
 
     @FXML 
-    TableView<Map.Entry<Integer, ProgramStatement>> codeTable;
+    TableView<Map.Entry<Long, ProgramStatement>> codeTable;
     @FXML
-    TableColumn<Map.Entry<Integer, ProgramStatement>, String> codeAddressColumn;
+    TableColumn<Map.Entry<Long, ProgramStatement>, String> codeAddressColumn;
     @FXML
-    TableColumn<Map.Entry<Integer, ProgramStatement>, String> machineCodeColumn;
+    TableColumn<Map.Entry<Long, ProgramStatement>, String> machineCodeColumn;
     @FXML
-    TableColumn<Map.Entry<Integer, ProgramStatement>, String> codeOriginalStringColumn;
+    TableColumn<Map.Entry<Long, ProgramStatement>, String> codeOriginalStringColumn;
 
     @FXML
     TextField startAddressTextField;
@@ -203,11 +203,12 @@ public class MemoryController implements MemoryObserver {
         Simulator.setMemoryController(this);
 
         codeAddressColumn.setComparator(new NumberComparator());
-        codeAddressColumn.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Map.Entry<Integer, ProgramStatement>,String>,ObservableValue<String>>() {
+        codeAddressColumn.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Map.Entry<Long, ProgramStatement>,String>,ObservableValue<String>>() {
 
             @Override
-            public ObservableValue<String> call(CellDataFeatures<Map.Entry<Integer, ProgramStatement>, String> param) {
+            public ObservableValue<String> call(CellDataFeatures<Map.Entry<Long, ProgramStatement>, String> param) {
                 if (displayAddressAsHex) {
+                    // Class Cast Exception Long to Int
                     StringBuilder str = new StringBuilder(
                             Long.toHexString(param.getValue().getKey()).toUpperCase());
                     str.insert(0, "0x");
@@ -219,19 +220,19 @@ public class MemoryController implements MemoryObserver {
             
         });
 
-        machineCodeColumn.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Map.Entry<Integer, ProgramStatement>,String>,ObservableValue<String>>() {
+        machineCodeColumn.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Map.Entry<Long, ProgramStatement>,String>,ObservableValue<String>>() {
 
             @Override
-            public ObservableValue<String> call(CellDataFeatures<Map.Entry<Integer, ProgramStatement>, String> param) {
+            public ObservableValue<String> call(CellDataFeatures<Map.Entry<Long, ProgramStatement>, String> param) {
                 return new SimpleStringProperty(param.getValue().getValue().getMachineCodeString());
             }
             
         });
 
-        codeOriginalStringColumn.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Map.Entry<Integer, ProgramStatement>,String>,ObservableValue<String>>() {
+        codeOriginalStringColumn.setCellValueFactory(new Callback<TableColumn.CellDataFeatures<Map.Entry<Long, ProgramStatement>,String>,ObservableValue<String>>() {
 
             @Override
-            public ObservableValue<String> call(CellDataFeatures<Map.Entry<Integer, ProgramStatement>, String> param) {
+            public ObservableValue<String> call(CellDataFeatures<Map.Entry<Long, ProgramStatement>, String> param) {
                 return new SimpleStringProperty(param.getValue().getValue().getSource());
             }
             
@@ -475,7 +476,7 @@ public class MemoryController implements MemoryObserver {
     }
 
     public void updateCode(ARMProgram program) {
-        ObservableList<Map.Entry<Integer, ProgramStatement>> items = FXCollections.observableArrayList(program.getProgramStatements().entrySet());
+        ObservableList<Map.Entry<Long, ProgramStatement>> items = FXCollections.observableArrayList(program.getProgramStatements().entrySet());
         codeTable.setItems(items);
     }
 }
